re organize components into sub folders for easy maintainability
